test/fixtures/wpt/resource-timing/worklet-rt-entries.https.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing Entry Sequence of Events for Worklets</title>
<link rel="help" href="https://w3c.github.io/resource-timing/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
    promise_test(async() => {
        const url = new URL("/worklets/resources/empty-worklet-script.js", location.href).toString();
        await CSS.paintWorklet.addModule(url);
        assert_equals(performance.getEntriesByName(url).length, 1);
    }, "Worklets should generate Resource Timing Entries");
</script>
</body>
</html>
